pyproject.toml
[project]
name = "tg-tools"
dynamic = ["version"]
description = "Ferramentas para userbots e bots no Telegram."
readme = "README.md"
authors = [
    { name = "Vitor", email = "vitorribeiro062@gmail.com" }
]
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "hydrogram>=0.2.0",
    "pathvalidate>=3.2.3",
    "pillow>=11.1.0",
    "pyfiglet>=1.0.2",
    "rich>=13.9.4",
    "tinydb>=4.8.2",
]

[project.scripts]
tg-tools = "tg_tools:main"

[project.optional-dependencies]
tgcrypto = ["tgcrypto"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/tg_tools/version.py"

[dependency-groups]
dev = ["pytest>=8.4.2", "pytest-cov>=7.0.0"]
